requirements.txt
requests
icecream
loguru
bs4
lxml
urllib3
fake-useragent
attrdict3
